Break unified_conn into many small modules
